// src/app/individual-lesson/model/validation-messages.model.ts
import { Observable } from 'rxjs';

export interface ValidationMessages {
  titleValidationMessage: Observable<string>;
  lessonDatesValidationMessage: Observable<string>;
  lessonStartDateValidationMessage: Observable<string>;
  lessonEndDateValidationMessage: Observable<string>;
  studentValidationMessage: Observable<string>;
  descriptionValidationMessage: Observable<string>;
  lessonTimesValidation: Observable<string>;
  weekdaysWithTimeRangesValidationMessage: Observable<string>;
  lessonsTitlesValidationMessage: Observable<string>;
  lessonsDurationValidationMessage: Observable<string>;
}